Remove hacks.
Now that lambency has the ability to render within a transformed context,
we don't need to carry around the yoffset to every wire that renders on
the board. As a result, we can take this requirement out of a bunch of
the wires functions that operate on the board. <3 Haskell.